refactor: lock aop 하드코딩 문자열 상수로 변경

+++ core/core-api/src/main/java/cmc/mellyserver/common/aspect/lock/OptimisticLock.java
@@ -9,7 +9,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface OptimisticLock {
 
-    long retryCount() default 5L;
+    int retryCount() default 5;
 
     long waitTime() default 1L;
 }
